Add token hit area utilities and remove drag-on-damage setting
