+++ lang/de.json
@@ -842,7 +842,6 @@
     "settingDamageMath": "Schaden: Zeige die Berechnung",
     "settingDamageRestrictADD": "Schaden: Beschränke Schadensdialog (ADD) auf SL",
     "settingDamageSimpleADD": "Schaden: Einfacher Schadensdialog (ADD)",
-    "settingDragOnTokenDamage": "Ziehe, um Schaden auf ausgewähltes Token anzuwenden",
     "settingEnhancedInput": "Akteur: Verbesserte numerische Eingaben",
     "settingFlagItems": "Akteur: Foundry Item Markierung anzeigen",
     "settingFlagUserCreated": "Akteur: Benutzer-erstellt Markierung anzeigen",
@@ -870,7 +869,6 @@
     "settingHintDamageMath": "Wenn diese Option aktiviert ist, wird der Abschnitt 'Zeige Berechnung' in der Schadens-Chat-Nachricht standardmäßig angezeigt. Wenn das Häkchen nicht gesetzt ist, kann auf den Titel 'Zeige Berechnung' geklickt werden, um ihn zu erweitern. HINWEIS: Wenn dies die letzte Nachricht im Chatprotokoll ist, wird sie möglicherweise unter dem unteren Rand des Protokolls aufgeklappt und man muß möglicherweise nach unten scrollen, um sie zu sehen.",
     "settingHintDamageRestrictADD": "Wenn diese Option aktiviert ist, kann nur die Spielleitung den Dialog 'Schadensdialog' sehen/benutzen. Spieler, die Schaden zuteilen, erhalten lediglich eine Nachricht.",
     "settingHintDamageSimpleADD": "Wenn diese Option aktiviert ist, wird nur die Option 'Direkt anwenden' im Schadensdialog angezeigt.",
-    "settingHintDragOnTokenDamage": "Wenn aktiviert, wird das Ziehen von Schaden aus dem Chatprotokoll auf ein Token den Schaden auf das Token anwenden. Wenn deaktiviert, wird ein Dialog angezeigt, um nach dem Ziel zu fragen.",
     "settingHintEnhancedInput": "Wenn diese Option aktiviert ist, werden in bestimmten Feldern (z. B. AP) Schaltflächen zum Addieren und Subtrahieren angezeigt, wenn sie ausgewählt sind.",
     "settingHintFlagItems": "Wenn diese Option aktiviert ist, erscheint ein kleines Symbol hinter Gegenständen (und Funktionen), die von Foundry-Items kommen.",
     "settingHintFlagUserCreated": "Wenn diese Option aktiviert ist, wird ein kleines Symbol hinter von Benutzer-erstellten (nicht importierten) Gegenständen und vor von Benutzer-erstellten Notizen angezeigt.",
