+++ lang/en.json
@@ -1099,7 +1099,6 @@
     "settingDamageRestrictADD": "Damage: Restrict ADD to GM",
     "settingDamageSimpleADD": "Damage: Simple ADD",
     "settingDefaultADDAction": "Set default action for Damage Dialog",
-    "settingDragOnTokenDamage": "On drag, apply damage to selected token",
     "settingEnhancedInput": "Actor: Enhanced numeric inputs",
     "settingFlagItems": "Actor: Display Foundry Item Flag",
     "settingFlagUserCreated": "Actor: Display User-created Flag",
@@ -1128,7 +1127,6 @@
     "settingHintDamageMath": "If checked, the 'SHOW THE MATH' section of the damage chat message will be expanded by default. If unchecked, you can click on the 'SHOW THE MATH' title to expand it. NOTE: If this is the last message in the chat log, it may expand below the bottom of the log, and you might need to scroll down to see it.",
     "settingHintDamageRestrictADD": "If checked, only GMs can view/use the Apply Damage Dialog. Players who drag and drop damage will just receive a message.",
     "settingHintDamageSimpleADD": "If checked, only display the 'Directly Apply' option in the Apply Damage Dialog.",
-    "settingHintDragOnTokenDamage": "If checked, dragging damage from the chat log to a token will apply the damage to the token. If unchecked, a dialog will show to ask for the target.",
     "settingHintefaultADDAction": "When set to 'Target' it will choose 'Apply' for Player Characters and 'Apply Quietly' for NPCs.",
     "settingHintEnhancedInput": "If checked, certain fields (such as FP) will display add and subtract shortcut buttons when they have focus.",
     "settingHintFlagItems": "If checked, a small icon will appear after equipment (and features) created from Foundry Items",
