+++ lang/fr.json
@@ -988,7 +988,6 @@
     "settingDamageMath": "Dégâts: Montrer les maths",
     "settingDamageRestrictADD": "Dégâts: Restreindre ADD au MJ",
     "settingDamageSimpleADD": "Dégâts: ADD Simple",
-    "settingDragOnTokenDamage": "Glisser pour appliquer les dégâts au token sélectionné",
     "settingEnhancedInput": "Acteur: Entrées numériques améliorées",
     "settingFlagItems": "Acteur: Afficher le Flag 'Objet Foundry'.",
     "settingFlagUserCreated": "Acteur: Afficher les Flags créés par les utilisateurs",
@@ -1016,7 +1015,6 @@
     "settingHintDamageMath": "Si coché, la section 'MONTRER LES MATHS' des messages de dégât du chat sera étendue par défaut. Si décoché, vous pouvez cliquer sur le titre 'MONTRER LES MATHS' pour l'étendre. NOTE: Si c'est le dernier message du chat, il peut s'étendre au delà du bas de la fenêtre et vous devrez faire défiler vers le bas pour le voir.",
     "settingHintDamageRestrictADD": "Si coché, seuls les MJs peuvent voir/utiliser la fenêtre d'Application Des Dégâts. Les joueurs qui glissent-déposent des dégâts recevront juste un message.",
     "settingHintDamageSimpleADD": "Si coché, affiche seulement l'option 'Appliquer Directement' dans la Fenêtre d'Application des Dégâts.",
-    "settingHintDragOnTokenDamage": "Si coché, glisser les dégâts du chat vers un token les appliquera au token. Si décoché, une boîte de dialogue demandera la cible.",
     "settingHintEnhancedInput": "Si coché, certain champs (comme les PF) verront s'ajouter et retirer des boutons raccourcis quand ils sont sélectionnés.",
     "settingHintFlagItems": "Si coché, une petite icône apparaîtra après les équipements (et traits) créés depuis des Objets Foundry",
     "settingHintFlagUserCreated": "Si coché, une petit icône apparaîtra après un équipement ou une note créé par l'utilisateur.",
