+++ lang/pt_br.json
@@ -1039,7 +1039,6 @@
     "settingDamageRestrictADD": "Dano: Restringir JAD para o MJ",
     "settingDamageSimpleADD": "Dano: JAD Simples",
     "settingDefaultADDAction": "Definir ação padrão para a Janela de Dano",
-    "settingDragOnTokenDamage": "Ao arrastar, aplicar dano ao token selecionado",
     "settingEnhancedInput": "Ator: Entradas numéricas melhoradas",
     "settingFlagItems": "Ator: Marcar Itens do Foundry",
     "settingFlagUserCreated": "Ator: Marcar criações do Usuário",
@@ -1067,7 +1066,6 @@
     "settingHintDamageMath": "Se marcado, a seção de 'EXIBIR CÁLCULOS' da mensagem de chat será, por padrão, expandida. Se desmarcado, você pode clicar no título 'EXIBIR CÁLCULOS' para expandi-lo. NOTA: Se ela for a última mensagem do chat, pode ser que ela se expanda para baixo do limite inferior da janela de chat, e pode ser que você precise rolar para baixo para visualizá-la.",
     "settingHintDamageRestrictADD": "Se marcado, só o MJ pode visualizar/utilizar a Janela de Avaliação de Dano. Jogadores de arrastarem e soltarem dano receberão uma mensagem.",
     "settingHintDamageSimpleADD": "Se marcado, só exibirá a opção de 'Aplicar Diretamente' na Janela de Avaliação de Dano.",
-    "settingHintDragOnTokenDamage": "Se marcado, arrastar dano do log de chat para um token aplicará o dano ao token. Se desmarcado, uma janela de diálogo será exibida para perguntar o alvo.",
     "settingHintefaultADDAction": "Quando configurado para 'Alvo' o sistema vai escolher 'Aplicar' para PdJs e 'Aplicar Discretamente' para PdMs.",
     "settingHintEnhancedInput": "Se marcado, certos campos (como PF) exibirão botões atalho de adição e subtração quando selecionados.",
     "settingHintFlagItems": "Se marcado, um pequeno ícone vai aparecer na frente de equipamento (e características) criadas a partir dos Itens do Foundry",
