+++ lang/ru.json
@@ -504,7 +504,6 @@
     "settingAutoAddTaggedModifiers": "Автоматически добавлять отмеченные модификаторы",
     "settingCheckConditionalModifiers": "Проверить условные модификаторы на листе актера",
     "settingDamageLocationTorso": "Торс",
-    "settingDragOnTokenDamage": "При перетаскивании применять урон к выбранному токену",
     "settingHintAddCumulativeParryPenalties": "Если включено, система автоматически добавит кумулятивные штрафы за парирование к броску для выбранных токенов для проверки максимальных действий.",
     "settingHintAddShockAtTurn": "Когда добавлять эффект шока на цели.",
     "settingHintAllowAfterMaxActions": "Если отмечено, система позволит игроку выполнить действие после достижения максимального количества действий за раунд.",
@@ -514,7 +513,6 @@
     "settingHintAnnounceCriticalRolls": "Если установлено, система будет объявлять, с анимацией и звуком, критические успехи и неудачи. Только для игроков и не скрытых бросков. Требуется установленный Sequencer.",
     "settingHintAutoAddTaggedModifiers": "Если установлено, система автоматически добавит отмеченные модификаторы к броску.",
     "settingHintCheckConditionalModifiers": "Если включено, система будет искать условные модификаторы на листе актера и добавлять их в всплывающее окно модификаторов эффектов.",
-    "settingHintDragOnTokenDamage": "Если установлено, перетаскивание урона из окна чата на токен применит урон к токену. Если не установлено, будет показан диалоговое окно для выбора цели.",
     "settingHintModifyDicePlusAdds": "Если установлено, система будет использовать правило 'Изменить кубики + Добавить' при добавлении модификаторов кубиков для бросков урона, как указано на стр. B269.",
     "settingHintShowConfirmationRoll": "Если установлено, система покажет диалог для подтверждения броска перед его выполнением.",
     "settingHintShowDebugInfo": "Для диалогов документов (Актёры, Предметы и т.д.) показывать значок отладки в заголовке окна. При нажатии он отобразит данные документа в диалоге.",
