+++ lib/miscellaneous-settings.js
@@ -77,7 +77,7 @@ export const SETTING_USE_FOUNDRY_ITEMS = 'use-foundry-items'
 export const SETTING_SHOW_DEBUG_INFO = 'show-debug-info'
 export const SETTING_SHOW_FOUNDRY_GLOBAL_ITEMS = 'show-foundry-global-items'
 export const SETTING_SHOW_ITEM_IMAGE = 'show-item-image'
-export const SETTING_DRAG_ON_TOKEN_DAMAGE = 'drag-on-token-damage'
+
 export const SETTING_USE_QUICK_ROLLS = 'use-quick-rolls'
 export const SETTING_SHOW_CONFIRMATION_ROLL_DIALOG = 'show-confirmation-roll-dialog'
 export const SETTING_ALLOW_ROLL_BASED_ON_MANEUVER = 'allow-roll-based-on-maneuver'
@@ -667,16 +667,6 @@ export function initializeSettings() {
       },
       default: 'target',
       onChange: value => console.log(`ADD apply option: ${value}`),
-    })
-
-    game.settings.register(SYSTEM_NAME, SETTING_DRAG_ON_TOKEN_DAMAGE, {
-      name: i18n('GURPS.settingDragOnTokenDamage'),
-      hint: i18n('GURPS.settingHintDragOnTokenDamage'),
-      scope: 'world',
-      config: true,
-      type: Boolean,
-      default: true,
-      onChange: value => console.log(`Drag on token damage : ${value}`),
     })
 
     // Status Effects Configuration ----
