+++ templates/apply-damage/select-token.hbs
@@ -5,6 +5,7 @@
       <input type='checkbox' name='all' value='all' />
       <b>{{i18n 'GURPS.allTokens'}}</b>
     </label>
+    <hr />
     {{#each tokens}}
       <label>
         <input type='checkbox' name='tokens' value='{{this.name}}' />
